Parte de paciente

+++ PI_Emi_Tania/PI_Emi_Tania/src/main/java/com/example/PI_Emi_Tania/controller/PacienteController.java
@@ -41,7 +41,7 @@ public class PacienteController {
     }
 
     @DeleteMapping("eliminar/{id}")
-    public void eliminar(@PathVariable int id) {
+    public void eliminar(@PathVariable Long id) {
         pacienteService.eliminarPorId(id);
     }
